Femininomenon
Same old story, time again
Got so close but then you lost it
Should've listened to your friends
About his girlfriend back in Boston
You sent him pictures and playlists and phone sex
He disappeared from the second that you said
"Let's get coffee, let's meet up"
I'm so sick of online love
And I don't understand, why can't any man
Hit it like
Get it hot
Make a bitch
It's a fem
Hit it like
Get it hot
Make a bitch
It's a-, um, can you play a song with a fucking beat?
Hit it like rom-pom-pom-pom (hit it like)
Get it hot like Papa John (get it hot)
Make a bitch go on and on (make a bitch)
It's a femininomenon (it's a fem)
Hit it like rom-pom-pom-pom (hit it like)
Get it hot like Papa John (get it hot)
Make a bitch go on and on (make a bitch)
It's a femininomenon (it's a fem)
So let's say it's working out
You pretend to love his mother
Lying to your friends about
How he's such a god damn good lover
Stuck in the suburbs, you're folding his laundry
Got what you wanted, so stop feeling sorry
Crying at the nail salon
I'm so sick of online love
I don't understand, why can't any man
Hit it like
Get it hot
Make a bitch
It's a fem
Hit it like
Get it hot
Make a bitch
It's a-, dude, can you play a song with a fucking beat?
Hit it like rom-pom-pom-pom (hit it like)
Get it hot like Papa John (get it hot)
Make a bitch go on and on (make a bitch)
It's a femininomenon (it's a fem)
Hit it like rom-pom-pom-pom (hit it like)
Get it hot like Papa John (get it hot)
Make a bitch go on and on (make a bitch)
It's a femininomenon (it's a fem)
Ladies, you know what I mean
And you know what you need and so does he
But does it happen? (No)
But does it happen? (No)
Well, what we really need is a femininomenon (a what?)
A femininomenon
Hit it like rom-pom-pom-pom (hit it like)
Get it hot like Papa John (get it hot)
Make a bitch go on and on (make a bitch)
It's a femininomenon (it's a fem)
Hit it like rom-pom-pom-pom (hit it like)
Get it hot like Papa John (get it hot)
Make a bitch go on and on (make a bitch)
Did you hear me? Play the fucking beat
Hit it like rom-pom-pom-pom (hit it like)
Get it hot like Papa John (get it hot)
Make a bitch go on and on (make a bitch)
It's a femininomenon (it's a fem)
Hit it like rom-pom-pom-pom (hit it like)
Get it hot like Papa John (get it hot)
Make a bitch go on and on (make a bitch)
It's a femininomenon (it's a fem)
Hit it like
Get it hot
Make a bitch
It's a fem (it's a femininomenon)
Hit it like
Get it hot
Make a bitch
It's a fem (fem, fem, fem)